Fix failing tests

+++ compiler/testData/codegen/bytecodeText/boxingOptimization/casts.kt
@@ -9,7 +9,7 @@ fun bar() {
     foo(1L) { x -> x!!.toByte() }
     foo(1L) { x -> x!!.toShort() }
     foo('a') { x -> x!!.toDouble() }
-    foo(1.0) { x -> x!!.toByte() }
+    foo(1.0) { x -> x!!.toInt() }
 }
 
 // 0 valueOf
